backend/requirements.txt
boto3==1.35.0
aws-cdk-lib==2.150.0
constructs>=10.0.0
python-dotenv==1.0.0
pydantic==2.8.0
